# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Channel Doctor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /channels/search:
    get:
      operationId: searchChannels
      tags: [channels]
      summary: Search YouTube for channels by query (handle, URL, or name)
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: Top results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSearchResults"
  /channels/{channelId}:
    get:
      operationId: getChannel
      tags: [channels]
      summary: Full channel snapshot (stats + recent videos + health)
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Channel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelDetails"
  /channels/{channelId}/analysis:
    get:
      operationId: analyzeChannel
      tags: [channels]
      summary: AI-powered diagnosis and growth strategy
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelAnalysis"
...
  /channels/{channelId}/content-plan:
    get:
      operationId: getContentPlan
      tags: [channels]
      summary: AI weekly upload schedule for a channel
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentPlanResult"
  /channels/{channelId}/forecast:
    get:
      operationId: getForecast
      tags: [channels]
      summary: 30-day growth forecast
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Forecast
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForecastResult"
...
  /channels/{channelId}/video-breakdown:
    get:
      operationId: getVideoBreakdown
      tags: [channels]
      summary: AI breakdown of recent videos (title score, hook, CTR estimate)
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Video breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoBreakdownResult"
  /channels/{channelId}/competitors:
    get:
      operationId: compareCompetitors
      tags: [channels]
      summary: Find similar channels and AI-compare
      parameters:
        - in: path
          name: channelId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Competitors comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitorsResult"
...
components:
  schemas:
...
    ChannelSearchResult:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        thumbnail: { type: string }
        description: { type: string }
        customUrl: { type: string, nullable: true }
      required: [id, title, thumbnail, description]
    ChannelSearchResults:
      type: object
      properties:
        results:
          type: array
          items: { $ref: "#/components/schemas/ChannelSearchResult" }
      required: [results]
    VideoSummary:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        thumbnail: { type: string }
        publishedAt: { type: string }
        durationSeconds: { type: integer }
        views: { type: integer }
        likes: { type: integer }
        comments: { type: integer }
      required: [id, title, thumbnail, publishedAt, durationSeconds, views, likes, comments]
    ViewTrendPoint:
      type: object
      properties:
        date: { type: string }
        views: { type: integer }
        videoId: { type: string }
      required: [date, views, videoId]
    UploadCadencePoint:
      type: object
      properties:
        weekStart: { type: string }
        uploads: { type: integer }
      required: [weekStart, uploads]
    BestPostingHour:
      type: object
      properties:
        dayOfWeek: { type: integer }
        hour: { type: integer }
      required: [dayOfWeek, hour]
    HealthBreakdown:
      type: object
      properties:
        engagement: { type: number }
        consistency: { type: number }
        growth: { type: number }
        performance: { type: number }
      required: [engagement, consistency, growth, performance]
    ChannelDetails:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        thumbnail: { type: string }
        bannerUrl: { type: string, nullable: true }
        country: { type: string, nullable: true }
        publishedAt: { type: string }
        subscriberCount: { type: integer }
        hiddenSubscriberCount: { type: boolean }
        viewCount: { type: integer }
        videoCount: { type: integer }
        avgViews: { type: integer }
        medianViews: { type: integer }
        uploadCadencePerWeek: { type: number }
        engagementRate: { type: number }
        growthRatio: { type: number }
        viewsPerSubRatio: { type: number }
        bestPostingHour: { $ref: "#/components/schemas/BestPostingHour" }
        healthScore: { type: number }
        healthStatus: { type: string, enum: [critical, warning, healthy, thriving] }
        healthBreakdown: { $ref: "#/components/schemas/HealthBreakdown" }
        recentVideos:
          type: array
          items: { $ref: "#/components/schemas/VideoSummary" }
        viewTrend:
          type: array
          items: { $ref: "#/components/schemas/ViewTrendPoint" }
        uploadCadence:
          type: array
          items: { $ref: "#/components/schemas/UploadCadencePoint" }
      required: [id, title, description, thumbnail, publishedAt, subscriberCount, hiddenSubscriberCount, viewCount, videoCount, avgViews, medianViews, uploadCadencePerWeek, engagementRate, growthRatio, viewsPerSubRatio, bestPostingHour, healthScore, healthStatus, healthBreakdown, recentVideos, viewTrend, uploadCadence]
    NextAction:
      type: object
      properties:
        priority: { type: string, enum: [high, medium, low] }
        action: { type: string }
        impact: { type: string }
      required: [priority, action, impact]
    ChannelAnalysis:
      type: object
      properties:
        channelId: { type: string }
        healthScore: { type: number }
        healthStatus: { type: string }
        diagnosis: { type: string }
        strengths:
          type: array
          items: { type: string }
        weaknesses:
          type: array
          items: { type: string }
        opportunities:
          type: array
          items: { type: string }
        nextActions:
          type: array
          items: { $ref: "#/components/schemas/NextAction" }
        contentNiche: { type: string }
        audienceInsight: { type: string }
      required: [channelId, healthScore, healthStatus, diagnosis, strengths, weaknesses, opportunities, nextActions, contentNiche, audienceInsight]
...
    VideoBreakdownItem:
      type: object
      properties:
        videoId: { type: string }
        title: { type: string }
        thumbnail: { type: string }
        views: { type: integer }
        publishedAt: { type: string }
        titleScore: { type: number }
        thumbnailCritique: { type: string }
        hookStrength: { type: string, enum: [weak, ok, strong] }
        ctrCategory: { type: string, enum: [low, average, high, viral] }
        verdict: { type: string }
      required: [videoId, title, thumbnail, views, publishedAt, titleScore, thumbnailCritique, hookStrength, ctrCategory, verdict]
    VideoBreakdownResult:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/VideoBreakdownItem" }
      required: [items]
    CompetitorChannel:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        thumbnail: { type: string }
        subscriberCount: { type: integer }
        viewCount: { type: integer }
        videoCount: { type: integer }
        avgViews: { type: integer }
        uploadsPerWeek: { type: number }
        engagementRate: { type: number }
      required: [id, title, thumbnail, subscriberCount, viewCount, videoCount, avgViews, uploadsPerWeek, engagementRate]
    CompetitorsResult:
      type: object
      properties:
        competitors:
          type: array
          items: { $ref: "#/components/schemas/CompetitorChannel" }
        comparison: { type: string }
        advantages:
          type: array
          items: { type: string }
        gaps:
          type: array
          items: { type: string }
      required: [competitors, comparison, advantages, gaps]
...
    PlanDay:
      type: object
      properties:
        day: { type: string }
        topic: { type: string }
        format: { type: string }
        hook: { type: string }
        why: { type: string }
      required: [day, topic, format, hook, why]
    ContentPlanResult:
      type: object
      properties:
        cadenceAdvice: { type: string }
        schedule:
          type: array
          items: { $ref: "#/components/schemas/PlanDay" }
      required: [cadenceAdvice, schedule]
    ForecastResult:
      type: object
      properties:
        projectedSubs30d: { type: integer }
        projectedViews30d: { type: integer }
        confidence: { type: string, enum: [low, medium, high] }
        summary: { type: string }
        drivers:
          type: array
          items: { type: string }
        risks:
          type: array
          items: { type: string }
      required: [projectedSubs30d, projectedViews30d, confidence, summary, drivers, risks]
...