# lib/api-zod/src/generated/types/viewTrendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Channel Doctor API
 * OpenAPI spec version: 0.1.0
 */

export interface ViewTrendPoint {
  date: string;
  views: number;
  videoId: string;
}